pyproject.toml
[project]
name = "resume-mate"
version = "0.1.0"
description = "AI-powered resume tailor"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "jinja2>=3.1.6",
    "litellm>=1.80.11",
    "playwright>=1.57.0",
    "pydantic>=2.12.5",
    "pypdf>=6.5.0",
    "python-docx>=1.2.0",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "typer>=0.21.0",
]

[project.scripts]
resume-mate = "resume_mate.main:app"

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade (modern syntax!)
]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
